# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * RobloxAI Agent API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface DeleteResult {
  success: boolean;
}

export interface Conversation {
  id: number;
  title: string;
  /** @nullable */
  mode?: string | null;
  /** @nullable */
  executor?: string | null;
  /** @nullable */
  gameTarget?: string | null;
  createdAt: string;
  updatedAt: string;
  messageCount: number;
}

export interface Message {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  /** @nullable */
  fileUrl?: string | null;
  /** @nullable */
  fileName?: string | null;
  createdAt: string;
}

export interface ConversationWithMessages {
  id: number;
  title: string;
  /** @nullable */
  mode?: string | null;
  /** @nullable */
  executor?: string | null;
  /** @nullable */
  gameTarget?: string | null;
  createdAt: string;
  updatedAt: string;
  messages: Message[];
}

export interface ConversationInput {
  title?: string;
  mode?: string;
  executor?: string;
  gameTarget?: string;
}

export interface ConversationUpdate {
  title?: string;
  mode?: string;
  executor?: string;
  gameTarget?: string;
}

export interface ConversationStats {
  totalConversations: number;
  totalMessages: number;
  totalScripts: number;
  recentConversations?: Conversation[];
}

export interface MessageInput {
  content: string;
  /** @nullable */
  fileUrl?: string | null;
  /** @nullable */
  fileName?: string | null;
  /** @nullable */
  mode?: string | null;
}

export interface ExtractedScript {
  index: number;
  code: string;
  language: string;
  messageId?: number;
}

export interface FileUploadInput {
  file?: Blob;
}

export interface UploadedFile {
  url: string;
  name: string;
  size: number;
  type: string;
}

export interface Settings {
  id: number;
  /** @nullable */
  executor?: string | null;
  /** @nullable */
  platform?: string | null;
  /** @nullable */
  obfuscation?: string | null;
  /** @nullable */
  robloxVersion?: string | null;
  /** @nullable */
  scriptStyle?: string | null;
  /** @nullable */
  defaultMode?: string | null;
  /** @nullable */
  systemPromptExtra?: string | null;
  /** @nullable */
  uiLibPreference?: string | null;
}

export interface SettingsInput {
  executor?: string;
  platform?: string;
  obfuscation?: string;
  robloxVersion?: string;
  scriptStyle?: string;
  defaultMode?: string;
  systemPromptExtra?: string;
  uiLibPreference?: string;
}

export interface ScriptAnalysisInput {
  code: string;
  /** @nullable */
  conversationId?: number | null;
}

export interface ScriptAnalysisResult {
  issues: string[];
  optimizations: string[];
  score: number;
  fixedCode: string;
  explanation: string;
}

export interface RemoteScanInput {
  code: string;
}

export interface RemoteEntry {
  name: string;
  type: string;
  path: string;
  args?: string[];
  /** @nullable */
  bypassSuggestion?: string | null;
}

export interface RemoteScanResult {
  remotes: RemoteEntry[];
  suggestions: string[];
}

export interface ErrorDiagnosticInput {
  errorLog: string;
  /** @nullable */
  originalCode?: string | null;
}

export interface ErrorDiagnosticResult {
  rootCause: string;
  fixedCode: string;
  explanation: string;
  /** @nullable */
  line: number | null;
}

export interface ProfilerInput {
  code: string;
}

export interface ProfilerIssue {
  type: string;
  description: string;
  /** @nullable */
  line: number | null;
  severity: string;
}

export interface ProfilerResult {
  issues: ProfilerIssue[];
  optimizedCode: string;
  fpsImpact: string;
  warnings: string[];
}

export interface GuiBuildInput {
  prompt: string;
  /** @nullable */
  imageBase64?: string | null;
  /** @nullable */
  style?: string | null;
}

export interface GuiBuildResult {
  guiJson: string;
  luaCode: string;
  preview: string;
}