# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: RobloxAI Agent API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /conversations/{id}/messages:
    get:
      operationId: listMessages
      tags: [messages]
      summary: List messages in a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
  /conversations/{id}/chat:
    post:
      operationId: sendMessage
      tags: [messages]
      summary: Send a message and get AI response
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "200":
          description: AI response message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
...
  /files/upload:
    post:
      operationId: uploadFile
      tags: [files]
      summary: Upload a file (max 7MB)
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/FileUploadInput"
      responses:
        "200":
          description: Uploaded file info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadedFile"
...
components:
  schemas:
...
    Message:
      type: object
      required: [id, conversationId, role, content, createdAt]
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        role:
          type: string
        content:
          type: string
        fileUrl:
          type: ["string", "null"]
        fileName:
          type: ["string", "null"]
        createdAt:
          type: string
    MessageInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
        fileUrl:
          type: ["string", "null"]
        fileName:
          type: ["string", "null"]
        mode:
          type: ["string", "null"]
...
    FileUploadInput:
      type: object
      properties:
        file:
          type: string
          format: binary
    UploadedFile:
      type: object
      required: [url, name, size, type]
      properties:
        url:
          type: string
        name:
          type: string
        size:
          type: integer
        type:
          type: string
...